feat: standardize 34 DSA curriculum problems, ensure 100% mathematical consistency across all student records and teams, and fix Vercel auth & meta tag

+++ backend/app/schemas/analytics.py
@@ -5,7 +5,7 @@ from app.schemas.team import TeamOut
 
 class TopicMasteryStats(BaseModel):
     topic: str
-    percentage: int
+    percentage: float
     total_solved: int
